renamed users namespace
