+++ API/SSW.Rewards.Application/Common/Extensions/UserExtensions.cs
@@ -1,4 +1,4 @@
-﻿using SSW.Rewards.Application.User.Queries.GetCurrentUser;
+﻿using SSW.Rewards.Application.Users.Queries.GetCurrentUser;
 using System.Net.Mail;
 
 namespace SSW.Rewards.Application.Common.Extensions
