+++ API/SSW.Rewards.Application/User/Commands/UploadProfilePic/AutoMapperProfile.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace SSW.Rewards.Application.User.Commands.UploadProfilePic
+namespace SSW.Rewards.Application.Users.Commands.UploadProfilePic
 {
     public class AutoMapperProfile : Profile
     {
