+++ API/SSW.Rewards.Application/User/Commands/UploadProfilePic/UploadProfilePicCommandValidator.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace SSW.Rewards.Application.User.Commands.UploadProfilePic
+namespace SSW.Rewards.Application.Users.Commands.UploadProfilePic
 {
     public class UploadProfilePicCommandValidator : AbstractValidator<UploadProfilePicCommand>
     {
