+++ API/SSW.Rewards.Application/User/Commands/UpsertUser/UpsertUserCommandValidator.cs
@@ -1,6 +1,6 @@
 using FluentValidation;
 
-namespace SSW.Rewards.Application.User.Commands.UpsertUser
+namespace SSW.Rewards.Application.Users.Commands.UpsertUser
 {
     public class UpsertUserCommandValidator : AbstractValidator<UpsertUserCommand>
     {
