+++ API/SSW.Rewards.Application/User/Queries/GetUserAchievements/UserAchievementViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace SSW.Rewards.Application.User.Queries.GetUserAchievements
+namespace SSW.Rewards.Application.Users.Queries.GetUserAchievements
 {
     public class UserAchievementViewModel
     {
