+++ API/SSW.Rewards.Application/User/Queries/GetUserAchievements/UserAchievementsViewModel.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace SSW.Rewards.Application.User.Queries.GetUserAchievements
+namespace SSW.Rewards.Application.Users.Queries.GetUserAchievements
 {
     public class UserAchievementsViewModel
     {
