+++ API/SSW.Rewards.Application/User/Queries/GetUserRewards/JoinedUserReward.cs
@@ -1,6 +1,6 @@
 ﻿using SSW.Rewards.Domain.Entities;
 
-namespace SSW.Rewards.Application.User.Queries.GetUserRewards
+namespace SSW.Rewards.Application.Users.Queries.GetUserRewards
 {
     internal class JoinedUserReward
     {
